// server/src/enums/permissions.ts
export enum permission {
    CREATE_MODULE = "CREATE_MODULE",
    CREATE_PERMISSION = "CREATE_PERMISSION",
    CREATE_ROLE = "CREATE_ROLE",
    CREATE_SESSION = "CREATE_SESSION",
    EDIT_MODULE = "EDIT_MODULE",
    EDIT_PERMISSION = "EDIT_PERMISSION",
    EDIT_ROLE = "EDIT_ROLE",
    EDIT_SESSION = "EDIT_SESSION",
    DELETE_MODULE = "DELETE_MODULE",
    DELETE_PERMISSION = "DELETE_PERMISSION",
    DELETE_ROLE = "DELETE_ROLE",
    DELETE_SESSION = "DELETE_SESSION"
}